+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added "configure" button to custom feed header.
 - Remove OAuth client credentials from local storage if they are not valid.
 - Increased line height in session table.
+- Replaced application version with hostname in OAuth client name.
 
 ## [4.16.0] - 2026-01-02